Add FileCaller for testing purposes
The FileCaller class can be used for testing using sample API responses.

It gets the information from a static YAML file with the basic fields
needed to fake a real API response.

You can use any of the TestMakerCaller classes to obtain a sample of the
result XML or JSON from an API call and feed it to the 'content' field
of the test YAML file.