SQS: Return single attributes
If requesting a single attribute of an SQS queue, the Paws::API::Caller
gets a hashref, but looks like the code expects always to have an
arrayref in the end.

Add the particular case of not being able to detect the fields that
should be returned as the arrayref as just converting the hashref to an
arrayref with just a single hashref item.

Fixes #147