Make valid signatures for S3-alike services on non-default ports
When trying to use Paws::S3 with an S3 compatible server (minio in this case)
which listens on a non-default http(s) port, invalid signatures are generated
because only the `host` part of the endpoint URI is used.

AWS and minio server side use `host:port` when the port isn't the
default port (443).

Confirmed behaviour with aws cli/boto and minio's python client.