Ignore provided 'region' if 'credentialScope' is defined
If we end in a rule with 'credentialScope', we probably want to ignore
the 'region' even if the user has defined one.

Request to rules with 'credentialScope' are only going to succeed with
the region in it, so if the user (mistakenly or not) has a 'region'
defined the request will fail.

It may be nice, though, to leave a WARNING if 'region' is defined and is
different from the one in 'credentialScope'.