Make 'Paws' dependency explicit on Paws::API::Caller
This class uses Paws->load_class internally but does not declare it with
a 'use'. Most of the time it's ok because you already have Paws loaded,
but if you try to load Paws::API::Caller alone and call one of it's
methods, it will fail.

Adding this 'use Paws' prevents this eventual problem, and also makes a
bit clearer what depends on what on the Paws ecosystem.