Add common util functions to Paws.pm
All three `new_with_coercions` and `to_hash` and `is_internal_type` are
called in multiple places both in Paws code and for clients using Paws.

Having them in Paws::API::Caller it's a bit awkward for me, because I
don't associate these utilities to the caller (which, as far as I
understand, is designed to allow caller pluggability).

I've considered creating a Paws::Util or something like that, but it
looked to me that `load_class`, that was already under Paws.pm, was the
same kind of utility, and therefore I've put the three of them together.