bin/paws should not add lib to @INC
bin/paws binary seem to be designed for end
users and should not add an unsafe directory
to @INC. As we cannot predict where will be
the 'auto-lib' and 'lib' directories.

bin/paws_make_testcase is a tool for the distribution
and should not be shipped (view #309), we can add lib
to it to preserve its current behavior.