Fix for issue #171.
Due to a bug (?) in the URI module, a query string without an "=" sign
is not considered a query string. Unfortunately, this causes $qparam to
 become empty (i.e. to delete the "versions" query parameter) at line #218
in lib/Paws/Net/RestXmlCaller.pm. Which will later turn the GET request
to the REST API into something other than a versions request, and the
returned response will not make sense.