caching class load and role checks
Turns out that Moose (and Moo fwiw) role checks are pretty slow and
for a given response like dynamodb, decoding it involves 100's or
1000's of these role checks as we generated constructor arguments for
all the nested objects.

Since all the roles in question here are on generated classes that we
know won't be needing new roles applied at runtime, we can just do the
lookup once and cache it for next time.  This _massively_ improves the
JSON decoder speed.  With the new benchmark/decode-benchmark script
the results are impressive:

Paws 0.43 + this patch: 11.3s
Paws 0.43:              33.4s

Note: this work is not mine and simply a minimal patch extracted from
the feature/mooification branch:

    commit 9e98a091850416db48db4168f1ac206841955ed3
    Author:     Jess Robinson <j.robinson@shadowcat.co.uk>
    AuthorDate: Tue 24 Sep 2019 12:29:11 PM PDT -0700

        More speed up improvements caching class load and role checks