fix dynamodb retries
DynamoDB became near unusable due to this commit:

    commit e3dda3a7aa9b37076967c3115d77253ef22445cf
    CommitDate: Tue 02 Jul 2019 11:27:39 AM PDT -0700

        DynamoDB docu update

This is actually because the "ProvisionedThroughputExceededException"
was lifted from being a service-specific retry exception (i.e. in the
dynamodb service json file) to a globally defined
exception (botocore/data/_retry.json).

For the case of dynamodb specifically,
ProvisionedThroughputExceededException is a regular occurrence
especially when using autoscaling so without this change, dynamodb is
practically unusable in Paws!

After this commit, Paws/API/Retry (and its list of default rules) will
be directly generated from botocore/data/_retry.json so hopefully we
won't hit something like this again.

POC script:

    #!/usr/bin/env perl

    use v5.34;
    use warnings;

    use Paws;
    my $paws = Paws->new(
      config => {
        region => 'us-west-2',
        caller => 'Paws::Net::FurlCaller',
      },
    );
    my $ddb = $paws->service('DynamoDB');

    my $i = 0;
    while (1) {
      $i++;
      eval {
        my $item =  $ddb->GetItem(
          TableName => 'low-rsu-table',
          Key => {
            cache_key => { S => 'foobar'.$i },
          }
        );
      };
      if ($@) {
        warn "died after $i items";
        die $@;
      }

    }

Where 'low-rsu-table' is a dynamodb table configured at a low RSU (I
had it set to 1 with autoscaling).  I ran 10 workers of this script
with a hack like: `seq 1 10 | xargs -n1 -P10 perl test-dynamodb.pl`

With 0.43 things started dying near immediately (because it wouldn't
retry on the throughput exceeded exception).

    died after 2 items at test-dynamodb.pl line 27.
    The level of configured provisioned throughput for the table was exceeded. Consider increasing your provisioning level with the UpdateTable API.

With this patch it retries like a champ (and I could basically just
leave it running as long as I wanted while seeing tons of throttling
exceptions in the AWS monitoring console).